+++ tests/integration/main.rs
@@ -9,6 +9,7 @@ mod common;
 
 mod cli_reporting;
 mod ecosystem_tier7;
+mod enable_wrapper;
 #[cfg(target_os = "linux")]
 mod env_stripping;
 mod git_hooks;